i'm not GroupMaster

+++ Fourcus/src/main/java/Main.java
@@ -103,18 +103,23 @@ public class Main {
     
     public static void groupManage(BufferedReader br) throws IOException{
         System.out.println("그룹 관리");
-        groupService.selectGroup(br);
+        boolean bl = groupService.selectGroup(br);
 
-        System.out.println("1.그룹명 수정 2.공지사항 3.그룹원 관리 4.그룹 삭제");
-        System.out.print("메뉴를 선택하세요 : ");
-        int select = Integer.parseInt(br.readLine());
+        if(bl) {
 
-        switch (select) {
-            case 1 -> groupService.updateGroupName(br);
-            case 2 -> groupService.Notice(br);
-            case 3 -> groupMemberManage(br);
-            case 4 -> groupService.deleteGroup(br);
-            default -> throw new IllegalStateException("Unexpected value: " + select);
+            System.out.println("1.그룹명 수정 2.공지사항 3.그룹원 관리 4.그룹 삭제");
+            System.out.print("메뉴를 선택하세요 : ");
+            int select = Integer.parseInt(br.readLine());
+
+            switch (select) {
+                case 1 -> groupService.updateGroupName(br);
+                case 2 -> groupService.Notice(br);
+                case 3 -> groupMemberManage(br);
+                case 4 -> groupService.deleteGroup(br);
+                default -> throw new IllegalStateException("Unexpected value: " + select);
+            }
+        } else{
+            group(br);
         }
     }
